requirements.txt
streamlit
youtube-transcript-api
yt-dlp
fastapi
uvicorn
google-generativeai
python-dotenv
